+++ govcd/api_vcd_test.go
@@ -193,6 +193,9 @@ type TestConfig struct {
 		Media           string `yaml:"mediaName,omitempty"`
 		PhotonOsOvaPath string `yaml:"photonOsOvaPath,omitempty"`
 	} `yaml:"media"`
+	Misc struct {
+		LdapContainer string `yaml:"ldapContainer,omitempty"`
+	} `yaml:"misc"`
 }
 
 // Test struct for vcloud-director.